pyproject.toml
[project]
name = "azdbr"
version = "0.1.0"
description = "Azure Databricks support library (experimental)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-identity>=1.19.0",
    "azure-storage-blob>=12.24.0",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.1",
]
